+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Agents system.
 - Conjunctions resolution agent.
+- Number agent (singular/plural relations) and related meaning.number module.
 - Coreference resolution in the parser through local fork of neuralcoref.
 - atom_gender(), atom_number(), atom_person(), atom_animacy() in parsers.
 - Coreference inference: gender, number and animacy.